+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/adapter/registro/AdapterRegistro.java
@@ -1,4 +1,4 @@
-package ajudamei.allan_arthur.com.ajuda_mei.adapter;
+package ajudamei.allan_arthur.com.ajuda_mei.adapter.registro;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -10,15 +10,15 @@ import android.widget.TextView;
 import java.util.List;
 
 import ajudamei.allan_arthur.com.ajuda_mei.R;
-import ajudamei.allan_arthur.com.ajuda_mei.domain.pessoa.Cliente;
+import ajudamei.allan_arthur.com.ajuda_mei.domain.registro.Registro;
 
 /**
- * Created by arthur on 03/07/17.
+ * Created by arthur on 28/06/17.
  */
 
-public class AdapterCliente extends ArrayAdapter<Cliente> {
+public class AdapterRegistro extends ArrayAdapter<Registro> {
 
-    public AdapterCliente(Context context, int resource, List<Cliente> objects) {
+    public AdapterRegistro(Context context, int resource, List<Registro> objects) {
         super(context, resource, objects);
     }
 
@@ -30,12 +30,12 @@ public class AdapterCliente extends ArrayAdapter<Cliente> {
             convertView = inflater.inflate(R.layout.dataregistro, parent, false);
         }
 
-        Cliente temp = getItem(position);
+        Registro temp = getItem(position);
 
         TextView data = (TextView) convertView.findViewById(R.id.txt_data);
-        data.setText("Nome: " + temp.getNome());
+        data.setText("Data: " + temp.getData());
         TextView qnt = (TextView) convertView.findViewById(R.id.txt_telefone);
-        qnt.setText("Telefone: " + temp.getTelefone());
+        qnt.setText("Qtd: " + String.valueOf(temp.getQuantidade()));
 
         return convertView;
     }
