+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/caixa/registro/pagamento/boleto/PagamentoBoletoActivity.java
@@ -11,7 +11,7 @@ import android.widget.Toast;
 import java.util.List;
 
 import ajudamei.allan_arthur.com.ajuda_mei.R;
-import ajudamei.allan_arthur.com.ajuda_mei.adapter.BoletoAdapter;
+import ajudamei.allan_arthur.com.ajuda_mei.adapter.registro.BoletoAdapter;
 import ajudamei.allan_arthur.com.ajuda_mei.caixa.registro.CaixaDaEmpresaActivity;
 import ajudamei.allan_arthur.com.ajuda_mei.database.DatabaseRegistroBoleto;
 import ajudamei.allan_arthur.com.ajuda_mei.domain.registro.Boleto;
