+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/caixa/registro/venda/produto_final/VendaProdutoActivity.java
@@ -10,7 +10,7 @@ import android.widget.Toast;
 
 import java.util.List;
 
-import ajudamei.allan_arthur.com.ajuda_mei.adapter.AdapterProduto;
+import ajudamei.allan_arthur.com.ajuda_mei.adapter.item.AdapterProduto;
 import ajudamei.allan_arthur.com.ajuda_mei.caixa.registro.CaixaDaEmpresaActivity;
 import ajudamei.allan_arthur.com.ajuda_mei.database.DatabaseProdutoFinal;
 import ajudamei.allan_arthur.com.ajuda_mei.domain.item.ItemProdutoFinal;
