+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/calculo/preco/venda/CalcularPrecoVendaEscolherActivity.java
@@ -9,7 +9,7 @@ import android.widget.ListView;
 
 import java.util.List;
 
-import ajudamei.allan_arthur.com.ajuda_mei.adapter.AdapterProduto;
+import ajudamei.allan_arthur.com.ajuda_mei.adapter.item.AdapterProduto;
 import ajudamei.allan_arthur.com.ajuda_mei.database.DatabaseProdutoFinal;
 import ajudamei.allan_arthur.com.ajuda_mei.MainActivity;
 import ajudamei.allan_arthur.com.ajuda_mei.R;
