+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/controle/estoque/produto_final/SelecionarMatPrimaActivity.java
@@ -15,7 +15,7 @@ import java.util.List;
 
 import ajudamei.allan_arthur.com.ajuda_mei.R;
 import ajudamei.allan_arthur.com.ajuda_mei.UsoGeral;
-import ajudamei.allan_arthur.com.ajuda_mei.adapter.AdapterMateria;
+import ajudamei.allan_arthur.com.ajuda_mei.adapter.item.AdapterMateria;
 import ajudamei.allan_arthur.com.ajuda_mei.database.DatabaseMateriaPrima;
 import ajudamei.allan_arthur.com.ajuda_mei.domain.item.ItemMateriaPrima;
