+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/controle/estoque/produto_final/ShowProdutoFinalActivity.java
@@ -12,7 +12,7 @@ import android.widget.Toast;
 
 import java.util.List;
 
-import ajudamei.allan_arthur.com.ajuda_mei.adapter.AdapterRegistro;
+import ajudamei.allan_arthur.com.ajuda_mei.adapter.registro.AdapterRegistro;
 import ajudamei.allan_arthur.com.ajuda_mei.database.DatabaseProdutoFinal;
 import ajudamei.allan_arthur.com.ajuda_mei.R;
 import ajudamei.allan_arthur.com.ajuda_mei.UsoGeral;
