+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/domain/registro/Venda.java
@@ -1,6 +1,4 @@
-package ajudamei.allan_arthur.com.ajuda_mei.caixa.registro.venda.produto_final;
-
-import ajudamei.allan_arthur.com.ajuda_mei.domain.registro.Registro;
+package ajudamei.allan_arthur.com.ajuda_mei.domain.registro;
 
 /**
  * Created by Allan on 22/06/2017.
