+++ app/src/main/java/ajudamei/allan_arthur/com/ajuda_mei/lista/ListaClientesActivity.java
@@ -15,7 +15,7 @@ import android.widget.Toast;
 import java.util.ArrayList;
 import java.util.List;
 
-import ajudamei.allan_arthur.com.ajuda_mei.adapter.AdapterCliente;
+import ajudamei.allan_arthur.com.ajuda_mei.adapter.pessoa.AdapterCliente;
 import ajudamei.allan_arthur.com.ajuda_mei.R;
 import ajudamei.allan_arthur.com.ajuda_mei.controle.estoque.materia_prima.ShowMatPrimaActivity;
 import ajudamei.allan_arthur.com.ajuda_mei.UsoGeral;
