+++ app/src/main/res/layout/activity_calcular_preco_venda_um.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ajudamei.allan_arthur.com.ajuda_mei.calculo.preco.venda.CalculoPrecoVendaUmActivity">
+    tools:context="ajudamei.allan_arthur.com.ajuda_mei.calculo.preco.venda.CalcularPrecoVendaUmActivity">
 
 
     <ImageView
